pyproject.toml
[project]
name = "proteinprediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.86",
    "goatools>=1.5.2",
    "jupyter>=1.1.1",
    "pandas>=2.3.3",
    "profun",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
profun = { git = "https://github.com/SamusRam/ProFun.git" }
